lesson2/seminar/step4_5_datahub_integration/requirements.txt
pandas~=2.2
scikit-learn~=1.5
dvc~=3.50
great-expectations~=0.18
pyyaml~=6.0
acryl-datahub~=0.12
requests~=2.31
